requirements.txt
Django==1.4
MySQL-python==1.2.3
South==0.7.5
cssselect==0.7.1
django-appconf==0.5
django-compressor==1.2
django-json-field==0.4.2
ipdb==0.6.1
ipython==0.12.1
lxml==3.1beta1
pyquery==1.2.4
python-dateutil==2.1
requests==1.1.0
six==1.2.0
wsgiref==0.1.2